+++ activestorage/lib/active_storage/service/gcs_service.rb
@@ -104,8 +104,10 @@ module ActiveStorage
       end
     end
 
-    def headers_for_direct_upload(key, checksum:, **)
-      { "Content-MD5" => checksum }
+    def headers_for_direct_upload(key, checksum:, filename: nil, disposition: nil, **)
+      content_disposition = content_disposition_with(type: disposition, filename: filename) if filename
+
+      { "Content-MD5" => checksum, "Content-Disposition" => content_disposition }
     end
 
     private